disable bpf_object__pin for now
This works.. the first time you run it.

If ran again, bpf_prog_load creates new maps and then the pinning
fails because the old ones are still pinned.

bpf_obj_get can be used to get a fd pointing to a pinned map, but when
bpf_prog_load runs, it doesn't know to check for existing maps.